refactor remote-tool
