+++ erniebot-agent/erniebot_agent/tools/base.py
@@ -14,13 +14,31 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from dataclasses import asdict, dataclass
 from typing import Any, List, Optional, Type
 
 from erniebot_agent.message import Message
-from erniebot_agent.tools.schema import ToolParameterView, scrub_dict
+from erniebot_agent.tools.schema import (
+    Endpoint,
+    EndpointInfo,
+    RemoteToolView,
+    ToolParameterView,
+    scrub_dict,
+)
+from yaml import safe_dump
 
 
-class Tool(ABC):
+class BaseTool(ABC):
+    @abstractmethod
+    async def __call__(self, *args: Any, **kwds: Any) -> Any:
+        raise NotImplementedError
+
+    @abstractmethod
+    def function_call_schema(self) -> dict:
+        raise NotImplementedError
+
+
+class Tool(BaseTool, ABC):
     description: str
     name: Optional[str] = None
     input_type: Optional[Type[ToolParameterView]] = None
@@ -55,3 +73,86 @@ class Tool(ABC):
     @property
     def examples(self) -> List[Message]:
         return []
+
+
+@dataclass
+class RemoteToolkit(BaseTool):
+    """plugin schema object which be converted from Toolkit and generate openapi configuration file"""
+
+    openapi: str
+    info: EndpointInfo
+    servers: List[Endpoint]
+    paths: List[RemoteToolView]
+
+    component_schemas: dict[str, Type[ToolParameterView]]
+
+    def to_openapi_dict(self) -> dict:
+        """convert plugin schema to openapi spec dict"""
+        spec_dict = {
+            "openapi": self.openapi,
+            "info": asdict(self.info),
+            "servers": [asdict(server) for server in self.servers],
+            "paths": {tool_view.uri: tool_view.to_openapi_dict() for tool_view in self.paths},
+            "components": {
+                "schemas": {
+                    uri: parameters_view.to_openapi_dict()
+                    for uri, parameters_view in self.component_schemas.items()
+                }
+            },
+        }
+        return scrub_dict(spec_dict, remove_empty_dict=True) or {}
+
+    def to_openapi_file(self, file: str):
+        """generate openapi configuration file
+
+        Args:
+            file (str): the path of the openapi yaml file
+        """
+        spec_dict = self.to_openapi_dict()
+        with open(file, "w+", encoding="utf-8") as f:
+            safe_dump(spec_dict, f, indent=4)
+
+    @staticmethod
+    def from_openapi_file(file: str) -> RemoteToolkit:
+        """only support openapi v3.0.1
+
+        Args:
+            file (str): the path of openapi yaml file
+        """
+        from openapi_spec_validator import validate
+        from openapi_spec_validator.readers import read_from_filename
+
+        spec_dict, _ = read_from_filename(file)
+        validate(spec_dict)
+
+        # info
+        info = EndpointInfo(**spec_dict["info"])
+        servers = [Endpoint(**server) for server in spec_dict.get("servers", [])]
+
+        # components
+        component_schemas = spec_dict["components"]["schemas"]
+        fields = {}
+        for name, schema in component_schemas.items():
+            parameter_view = ToolParameterView.from_openapi_dict(name, schema)
+            fields[name] = parameter_view
+
+        # paths
+        paths = []
+        for path, path_info in spec_dict.get("paths", {}).items():
+            for method, path_method_info in path_info.items():
+                paths.append(
+                    RemoteToolView.from_openapi_dict(
+                        uri=path,
+                        method=method,
+                        path_info=path_method_info,
+                        parameters_views=fields,
+                    )
+                )
+
+        return RemoteToolkit(
+            openapi=spec_dict["openapi"],
+            info=info,
+            servers=servers,
+            paths=paths,
+            component_schemas=fields,
+        )  # type: ignore
